# APM: Improve performance of QuantizeIP (APMSP-1986)