--- a/releasenotes/notes/ImproveQuantizePeerIPAddressesPerformance-5902d8baa33cda4a.yaml
+++ b/releasenotes/notes/ImproveQuantizePeerIPAddressesPerformance-5902d8baa33cda4a.yaml
@@ -1,0 +1,11 @@
+# Each section from every release note are combined when the
+# CHANGELOG.rst is rendered. So the text needs to be worded so that
+# it does not depend on any information only available in another
+# section. This may mean repeating some details, but each section
+# must be readable independently of the other.
+#
+# Each section note must be formatted as reStructuredText.
+---
+enhancements:
+  - |
+    APM: Improve the performance of the trace-agent's QuantizePeerIPAddresses function, this should result in a marginal reduction in CPU usage for most workloads.