<!--
* Contributors are encouraged to read our [CONTRIBUTING](/CONTRIBUTING.md) documentation.
* Both Contributor and Reviewer Checklists are available at https://datadoghq.dev/datadog-agent/guidelines/contributing/#pull-requests.
* The pull request:
  * Should only fix one issue or add one feature at a time.
  * Must update the test suite for the relevant functionality.
  * Should pass all status checks before being reviewed or merged.
* Commit titles should be prefixed with general area of pull request's change.
* Please fill the below sections if possible with relevant information or links.
-->
### What does this PR do?
Reduce CPU time and allocations of the QuantizeIPAddresses function. (Spotted by @paullegranddc )

Primarily there are 2 main optimizations here (viewable as separate commits:
1. Remove regexes and instead use strings.Index in a short loop
2. Do not use net.SplitHostPort as it allocates an error when the incoming value isn't parseable, (and we do not care about the exact error just that it exists).

I also investigated moving from strings.Split to using string.Cut and a string builder however it was not a clear-cut improvement. In the case where there is no comma in the incoming string it worsens performance.

### Motivation
A non-zero number of object allocations are done by this code and it is run against for all spans that are used to calculate stats (thus this happens before sampling and so there is a larger impact here compared to post-sampling features).


### Describe how you validated your changes
<!--
Validate your changes before merge, ensuring that:
* Your PR is tested by static / unit / integrations / e2e tests
* Your PR description details which e2e tests cover your changes, if any
* The PR description contains details of how you validated your changes. If you validated changes manually and not through automated tests, add context on why automated tests did not fit your changes validation.

If you want additional validation by a second person, you can ask reviewers to do it. Describe how to set up an environment for manual tests in the PR description. Manual validation is expected to happen on every commit before merge.

Any manual validation step should then map to an automated test. Manual validation should not substitute automation, minus exceptions not supported by test tooling yet.
-->
Existing unit tests have good coverage, to ensure no panics were inadvertently introduced I also added and ran a fuzz test.


I then ran and compared benchmarks using benchstat:
Comparing the regular expression changes shows a large CPU improvement when a match is found and removes an allocation:
```
goos: darwin
goarch: arm64
pkg: github.com/DataDog/datadog-agent/pkg/obfuscate
cpu: Apple M1 Max
                            │ prefixold.out │             prefix2.out             │
                            │    sec/op     │   sec/op     vs base                │
SplitPrefix/matching-10        354.90n ± 0%   13.86n ± 1%  -96.09% (p=0.000 n=10)
SplitPrefix/not_matching-10     37.08n ± 1%   37.26n ± 1%        ~ (p=0.147 n=10)
geomean                         114.7n        22.72n       -80.19%

                            │ prefixold.out │               prefix2.out               │
                            │     B/op      │    B/op     vs base                     │
SplitPrefix/matching-10        32.00 ± 0%      0.00 ± 0%  -100.00% (p=0.000 n=10)
SplitPrefix/not_matching-10    0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
geomean                                   ²               ?                       ² ³
¹ all samples are equal
² summaries must be >0 to compute geomean
³ ratios must be >0 to compute geomean

                            │ prefixold.out │               prefix2.out               │
                            │   allocs/op   │ allocs/op   vs base                     │
SplitPrefix/matching-10        1.000 ± 0%     0.000 ± 0%  -100.00% (p=0.000 n=10)
SplitPrefix/not_matching-10    0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
```

Then the larger  benchmark comparing using net split instead of our no-error-alloc version shows we reduce allocations when the incoming value isn't a "host:port".
```
goos: darwin
goarch: arm64
pkg: github.com/DataDog/datadog-agent/pkg/obfuscate
cpu: Apple M1 Max
                           │ useNetSplit.out │          ReplaceParse.out          │
                           │     sec/op      │   sec/op     vs base               │
QuantizePeerIpAddresses-10       14.84µ ± 2%   13.94µ ± 0%  -6.05% (p=0.000 n=10)

                           │ useNetSplit.out │           ReplaceParse.out           │
                           │      B/op       │     B/op      vs base                │
QuantizePeerIpAddresses-10      3.492Ki ± 0%   2.617Ki ± 0%  -25.06% (p=0.000 n=10)

                           │ useNetSplit.out │          ReplaceParse.out          │
                           │    allocs/op    │ allocs/op   vs base                │
QuantizePeerIpAddresses-10        150.0 ± 0%   122.0 ± 0%  -18.67% (p=0.000 n=10)
```


### Possible Drawbacks / Trade-offs

### Additional Notes
<!--
* Anything else we should know when reviewing?
* Include benchmarking information here whenever possible.
* Include info about alternatives that were considered and why the proposed
  version was chosen.
-->